requirements.txt
apex==0.1
fairseq==0.10.0
faiss==1.5.3
matplotlib==3.5.1
numpy==1.20.3
Pillow==9.1.1
scikit-learn==1.1.1
scipy==1.6.3
seaborn==0.11.1
SoundFile==0.10.3.post1
torch==1.10.0
torchvision==0.9.0
tqdm==4.60.0